Reorganized NRT tests.
